# OpenMP Parallelize InitDistribution