--- a/examples/cfchannel/analysis_cfchannel_10nC.py
+++ b/examples/cfchannel/analysis_cfchannel_10nC.py
@@ -79,7 +79,7 @@ print(
 )
 
 atol = 0.0  # ignored
-rtol = 1.5 * num_particles**-0.5  # from random sampling of a smooth distribution
+rtol = 2.1 * num_particles**-0.5  # from random sampling of a smooth distribution
 print(f"  rtol={rtol} (ignored: atol~={atol})")
 
 assert np.allclose(