--- a/examples/chicane/analysis_chicane_isr.py
+++ b/examples/chicane/analysis_chicane_isr.py
@@ -79,7 +79,7 @@ print(
 )
 
 atol = 0.0  # ignored
-rtol = 2.3 * num_particles**-0.5  # from random sampling of a smooth distribution
+rtol = 2.4 * num_particles**-0.5  # from random sampling of a smooth distribution
 print(f"  rtol={rtol} (ignored: atol~={atol})")
 
 assert np.allclose(