--- a/examples/epac2004_benchmarks/analysis_fodo_rf_SC.py
+++ b/examples/epac2004_benchmarks/analysis_fodo_rf_SC.py
@@ -79,7 +79,7 @@ sigx, sigy, sigt, emittance_x, emittance_y, emittance_t = get_moments(final)
 print(f"  sigx={sigx:e} sigy={sigy:e} sigt={sigt:e}")
 
 atol = 0.0  # ignored
-rtol = 3.5 * num_particles**-0.5  # from random sampling of a smooth distribution
+rtol = 3.6 * num_particles**-0.5  # from random sampling of a smooth distribution
 print(f"  rtol={rtol} (ignored: atol~={atol})")
 
 assert np.allclose(