--- a/examples/kurth/analysis_kurth_periodic.py
+++ b/examples/kurth/analysis_kurth_periodic.py
@@ -52,7 +52,7 @@ print(
 )
 
 atol = 0.0  # ignored
-rtol = 1.7 * num_particles**-0.5  # from random sampling of a smooth distribution
+rtol = 1.9 * num_particles**-0.5  # from random sampling of a smooth distribution
 print(f"  rtol={rtol} (ignored: atol~={atol})")
 
 assert np.allclose(
@@ -79,7 +79,7 @@ print(
 )
 
 atol = 0.0  # ignored
-rtol = 1.7 * num_particles**-0.5  # from random sampling of a smooth distribution
+rtol = 1.9 * num_particles**-0.5  # from random sampling of a smooth distribution
 print(f"  rtol={rtol} (ignored: atol~={atol})")
 
 assert np.allclose(