--- a/examples/optimize_triplet/analysis_triplet.py
+++ b/examples/optimize_triplet/analysis_triplet.py
@@ -80,7 +80,7 @@ print(
 )
 
 atol = 0.0  # ignored
-rtol = 2.0 * num_particles**-0.5  # from random sampling of a smooth distribution
+rtol = 2.5 * num_particles**-0.5  # from random sampling of a smooth distribution
 print(f"  rtol={rtol} (ignored: atol~={atol})")
 
 assert np.allclose(