--- a/examples/solenoid_restart/analysis_solenoid.py
+++ b/examples/solenoid_restart/analysis_solenoid.py
@@ -53,7 +53,7 @@ print(
 )
 
 atol = 0.0  # ignored
-rtol = 1.3 * num_particles**-0.5  # from random sampling of a smooth distribution
+rtol = 1.4 * num_particles**-0.5  # from random sampling of a smooth distribution
 print(f"  rtol={rtol} (ignored: atol~={atol})")
 
 assert np.allclose(