@ax3l 2025-07-30
I will push one more commit that slightly relaxes the tolerances for noise from having now slightly different init numbers when running with thread.

Individual threads on a process still draw their RNG distributions independently and with different seeds.